# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Label Manager API
  version: 0.1.0
  description: |
    REST API for the Label Manager application.

    This spec is the source of truth for the contract between frontend and backend, and now
    covers the whole served surface rather than the three endpoints the React client happens
    to call. Nothing enforces that yet: the conformance test that boots the app, reads
    `/v3/api-docs` and fails on any drift is Phase 3. Until it exists, an endpoint added
    without a change here will go unnoticed.

    Errors are RFC 9457 `ProblemDetail`, served as `application/problem+json`. Every path
    below also answers 401 when the session cookie is missing and 403 when a mutating request
    arrives without a valid CSRF token; neither is repeated per operation.

    Ownership: a nested id is checked against the label in its path, so asking for one label's
    release under another's id is a 404. That is not a tenant guard — nothing yet checks that
    the caller owns the label at all (F2, Phase 5).
paths:
...
  /api/labels/{labelId}/releases/{releaseId}/production-runs:
    parameters:
      - $ref: '#/components/parameters/LabelId'
      - $ref: '#/components/parameters/ReleaseId'
    get:
      operationId: listProductionRuns
      summary: The release's production runs, with inventory and movement history
      responses:
        '200':
          description: The production runs, possibly empty.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductionRunWithAllocation'
        '404':
          description: No such release under this label.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      operationId: createProductionRun
      summary: Record a pressing of this release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionRunRequest'
      responses:
        '201':
          description: Production run created. The response has no body.
        '404':
          $ref: '#/components/responses/NotFound'
  /api/labels/{labelId}/releases/{releaseId}/production-runs/{productionRunId}:
    parameters:
      - $ref: '#/components/parameters/LabelId'
      - $ref: '#/components/parameters/ReleaseId'
      - name: productionRunId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    delete:
      operationId: deleteProductionRun
      summary: Delete a production run
      responses:
        '204':
          description: Production run deleted.
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/labels/{labelId}/distributors:
    parameters:
      - $ref: '#/components/parameters/LabelId'
    get:
      operationId: listDistributors
      summary: The label's distributors
      responses:
        '200':
          description: The label's distributors, possibly empty.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Distributor'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createDistributor
      summary: Add a distributor to this label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, channelType]
              properties:
                name:
                  type: string
                channelType:
                  $ref: '#/components/schemas/ChannelType'
      responses:
        '201':
          description: Distributor created. The response has no body.
        '404':
          $ref: '#/components/responses/NotFound'
  /api/labels/{labelId}/distributors/{distributorId}:
    parameters:
      - $ref: '#/components/parameters/LabelId'
      - $ref: '#/components/parameters/DistributorId'
    get:
      operationId: getDistributor
      summary: Get one distributor
      description: |
        The distributor alone. Its sales, returns and agreements are sub-resources.
      responses:
        '200':
          description: The distributor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Distributor'
        '404':
          description: No such distributor under this label.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      operationId: deleteDistributor
      summary: Delete a distributor
      responses:
        '204':
          description: Distributor deleted.
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/labels/{labelId}/distributors/{distributorId}/agreements:
    parameters:
      - $ref: '#/components/parameters/LabelId'
      - $ref: '#/components/parameters/DistributorId'
    get:
      operationId: listAgreements
      summary: The distributor's pricing agreements
      description: |
        Each carries `productionRunId`. Resolve it against
        `/api/labels/{labelId}/production-runs` rather than expecting a rendered name.
      responses:
        '200':
          description: The agreements, possibly empty.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PricingAgreement'
        '404':
          description: No such distributor under this label.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      operationId: createAgreement
      summary: Price one production run for this distributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgreementRequest'
      responses:
        '201':
          description: Agreement created. The response has no body.
        '400':
          description: This distributor already has an agreement for that production run.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    LabelId:
      name: labelId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    ReleaseId:
      name: releaseId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    DistributorId:
      name: distributorId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    ProblemDetail:
      description: |
        RFC 9457 problem details, served as `application/problem+json`.

        Used by every error this application raises deliberately: the security filter
        chain's 401 and 403, and anything `ApiExceptionHandler` maps. An unmapped
        exception still falls through to Spring Boot's `/error` controller, which emits
        `{timestamp, status, error, path}` as `application/json` — that is a bug to be
        closed, not a second documented shape.
      type: object
      required:
        - status
      properties:
        type:
          type: string
          format: uri
          default: 'about:blank'
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
...
    Format:
      type: string
      enum: [DIGITAL, VINYL, CD, CASSETTE]
    ChannelType:
      type: string
      enum: [DIRECT, DISTRIBUTOR, RECORD_STORE, EVENT]
    CommissionType:
      type: string
      enum: [PERCENTAGE, FIXED_AMOUNT]
    LocationType:
      type: string
      enum: [WAREHOUSE, DISTRIBUTOR, EXTERNAL, BANDCAMP]
    MovementType:
      type: string
      enum: [ALLOCATION, SALE, RETURN, TRANSFER_IN, TRANSFER_OUT, ADJUSTMENT]
...
    ProductionRun:
      type: object
      required: [id, releaseId, format, quantity]
      properties:
        id: { type: integer, format: int64 }
        releaseId: { type: integer, format: int64 }
        format:
          $ref: '#/components/schemas/Format'
        description: { type: string, nullable: true }
        manufacturer: { type: string, nullable: true }
        manufacturingDate: { type: string, format: date, nullable: true }
        quantity: { type: integer }
    ProductionRunRequest:
      type: object
      required: [format, quantity]
      properties:
        format:
          $ref: '#/components/schemas/Format'
        description: { type: string, nullable: true }
        manufacturer: { type: string, nullable: true }
        manufacturingDate: { type: string, format: date, nullable: true }
        quantity: { type: integer }
    DistributorInventoryView:
      type: object
      required: [distributorId, current]
      properties:
        distributorId:
          type: integer
          format: int64
        current: { type: integer }
    MovementLocation:
      description: |
        Where stock moved from or to. `distributorId` is null unless the type is DISTRIBUTOR;
        join against `/api/labels/{labelId}/distributors` for the name.
      type: object
      required: [type]
      properties:
        type:
          $ref: '#/components/schemas/LocationType'
        distributorId:
          type: integer
          format: int64
          nullable: true
    MovementHistoryView:
      type: object
      required: [occurredAt, movementType, fromLocation, toLocation, quantity]
      properties:
        occurredAt: { type: string, format: date-time }
        movementType:
          $ref: '#/components/schemas/MovementType'
        fromLocation:
          $ref: '#/components/schemas/MovementLocation'
        toLocation:
          $ref: '#/components/schemas/MovementLocation'
        quantity: { type: integer }
    ProductionRunWithAllocation:
      type: object
      required: [productionRun, bandcampInventory, warehouseInventory]
      properties:
        productionRun:
          $ref: '#/components/schemas/ProductionRun'
        bandcampInventory: { type: integer }
        warehouseInventory: { type: integer }
        distributorInventories:
          type: array
          items:
            $ref: '#/components/schemas/DistributorInventoryView'
        movements:
          type: array
          items:
            $ref: '#/components/schemas/MovementHistoryView'
    Distributor:
      type: object
      required: [id, labelId, name, channelType]
      properties:
        id: { type: integer, format: int64 }
        labelId: { type: integer, format: int64 }
        name: { type: string }
        channelType:
          $ref: '#/components/schemas/ChannelType'
    PricingAgreement:
      type: object
      required: [id, distributorId, productionRunId, unitPrice, commissionType, commissionValue]
      properties:
        id: { type: integer, format: int64 }
        distributorId: { type: integer, format: int64 }
        productionRunId: { type: integer, format: int64 }
        unitPrice: { type: number }
        commissionType:
          $ref: '#/components/schemas/CommissionType'
        commissionValue: { type: number }
        createdAt: { type: string, format: date-time, nullable: true }
    AgreementRequest:
      type: object
      required: [productionRunId, unitPrice, commissionType, commissionValue]
      properties:
        productionRunId: { type: integer, format: int64 }
        unitPrice: { type: number }
        commissionType:
          $ref: '#/components/schemas/CommissionType'
        commissionValue: { type: number }
...
security:
  - cookieAuth: []